ai/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
numpy>=1.23
opencv-python>=4.7
pillow>=9.4
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
diffusers>=0.27.1
ultralytics>=8.0
python-multipart
accelerate
spotify

# py -3.10 -m venv testv
# .\testv\Scripts\activate
# pip install -r  .\requirements.txt
# uvicorn main:app --reload --host 0.0.0.0 --port 8080
